+++ INSTALLER-DARK-MODE/gradle/wrapper/gradle-wrapper.properties
