+++ INSTALLER-DARK-MODE/src/main/resources/ignore_test.png
