--- INSTALLER/build.gradle
@@ -1,29 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'tux.cool.is.loader'
-version '1.0'
-archivesBaseName = 'loader'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
-}
-
-test {
-    useJUnitPlatform()
-}
-jar {
-    manifest {
-        attributes "Main-Class": "tux.cool.is.loader.LoginGUI"
-    }
-
-    from {
-        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-    }
-}
