+++ README.md
@@ -1,2 +1,4 @@
 # InstallerForJava
 Lightweight installer written in java, made for minecraft mods, The installer uses JPanel and uses a URL to install to the specific area (Discord URL's work the best i've seen)
+
+Dark mode and light mode options are now available
